:recycle: 타이머 로직 useTimer 커스텀 훅으로 분리
